requirements.txt
langsmith>=0.1.39
python-dotenv>=1.0.1